Bugfix for handling of shadow doorbell in NVMe driver.
This patch adds full memory barrier into nvme_dbbuf_update_and_check_event function to ensure that the shadow doorbell is written before reading EventIdx from memory. This is a critical bugfix for initial patch that added support for shadow doorbell into NVMe driver:

https://github.com/torvalds/linux/commit/f9f38e33389c019ec880f6825119c94867c1fde0

This memory barrier is required as there is a possibility of out-of-order execution of two following operations in current driver implementation:

 - Write shadow doorbell (dbbuf_db) into memory.
 - Read EventIdx (dbbuf_ei) from memory.

When running in virtualized environment with shadow doorbell support the following logical sequence of events can occur:
 - Vm host processes all entries in NVMe queue.
 - Vm host updates EventIdx with last processed command Id.
 - Vm guest reads EventIdx (out of order execution).
 - Vm host reads shadow doorbell, and after comparing it with what was already processed decides to wait for regular MMIO doorbell from guest, as it looks like there is nothing to do.
 - Vm guest writes shadow doorbell (out of order execution).
 - Vm guest decides to not ring the MMIO doorbell as it looks like the (virtual) NVMe controller is still “behind” with processing commands.

(this sequence of events is not the only one that could trigger this bug)

This can result in hanging of virtual NVMe controller as guest won't issue MMIO doorbell on any of subsequent commands.

This issue is purely timing-dependent one, so there is no easy way to reproduce it. Our current approach is to run “ORacle IO Numbers” (ORION) that is shipped with Oracle DB:

./orion -run advanced -num_large 0 -size_small 8 -type rand -simulate concat -write 40 -duration 120 -matrix row -testname nvme_test

Where nvme_test is a .lun file that contains a list of NVMe block devices to run test against. Limiting number of vCPUs assigned to given VM instance seems to increase chances for this bug to occur (but there is no concrete proof for that). On test environment with VM that got 4 NVMe drives and 1 vCPU assigned we could observe virtual NVMe controller hang within 10-20 minutes. That correspond to about 400-500k IO operations processed (or about 100GB of IO read/writes).

We validated that the patch solves the issue by running Orion tool for 36 hours in a loop. This resulted in pushing 550M of IO operations (or about 150TB of IO read/writes) through the system.